Parsing content with go
